requirements.txt
transformers==4.49.0
torch==2.2.0+cu118
sentence-transformers==3.4.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-huggingface==0.1.2
pydantic==2.10.6
pydantic-settings==2.8.1
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
pytest==7.4.0
gunicorn==21.2.0
accelerate==1.5.2
bitsandbytes==0.45.3
chromadb==0.6.3
datasets==3.4.1
faiss-cpu==1.10.0
huggingface-hub==0.29.3
peft==0.15.1
safetensors==0.5.3
tokenizers==0.21.1
tiktoken==0.9.0
starlette==0.46.1
websockets==15.0.1 